% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrapml}
\alias{bootstrapml}
\title{arametric Bootstrap on Distributions Fitted with Maximum Likelihood}
\usage{
bootstrapml(obj, Nreps = 1000, map = identity,
  reducer = stats::quantile, ...)
}
\arguments{
\item{obj}{A \code{univariateML} object.}

\item{Nreps}{Positive integer. The number of bootstrap samples.}

\item{map}{A function of the parameters of the \code{univariateML} object.
Defaults to the identity.}

\item{reducer}{A reducer function. Defaults to \code{stats::quantile} with default
argument \code{probs = c(0.025, 0.975)}.}

\item{...}{Passed to reduce.}
}
\value{
The transposed map-reduced bootstrap samples.
}
\description{
arametric Bootstrap on Distributions Fitted with Maximum Likelihood
}
