% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-corbet.R
\docType{data}
\name{corbet}
\alias{corbet}
\title{Frequencies of butterflies collected in Malaya}
\format{
A vector of size 501 containing integer observations between 1 and 24.
}
\usage{
corbet
}
\description{
Species abundance data from 1943; a classical application of the logarithmic
series distribution.
}
\examples{
corbet
}
\references{
Fisher, R. A., Corbet, A. S., & Williams, C. B. (1943). The relation between the number of species and the number of individuals in a random sample of an animal population. The Journal of Animal Ecology, 12(1), 42. https://doi.org/10.2307/1411
}
\keyword{datasets}
