% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{decorator}
\alias{decorator}
\title{Decorate estimator}
\usage{
decorator(name)
}
\arguments{
\item{name}{Name of the \verb{ml***} function.}
}
\value{
The proper \verb{ml***} function.
}
\description{
The decorator constructs an \verb{ml***}. It first checks the input arguments,
then call the defined estimator (with elements \code{estimates} and \code{logLik}),
and the optional parameter \code{support}, then adds the required attributes.
}
\keyword{internal}
