% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-egypt.R
\docType{data}
\name{egypt}
\alias{egypt}
\title{Mortality data from ancient Egypt}
\format{A \link[tibble:tibble]{tibble} with 141 observations and 2 variables:
\describe{
\item{age}{Age at death.}
\item{sex}{Sex of deceased; 82 males and 49 females.}
}}
\usage{
egypt
}
\description{
Age at death of 141 Roman era Egyptian mummies.
}
\details{
This data was collected by Spiegelberg (1901) and analyzed by Karl Pearson
(1902) in the first volume of Biometrika. It was analyzed by
Claeskens & Hjort (2008) and the data is based on their transcription.
}
\examples{
egypt
}
\references{
Spiegelberg, W. (1901). Aegyptische und Griechische Eigennamen aus
Mumientiketten der Römischen Kaiserzeit.

Pearson, K. (1902). On the change in expectation of life in man during a
period of circa 2000 years. Biometrika, 1(2), 261-264.

Claeskens, G., & Hjort, N. L. (2008). Model selection and model averaging.
Cambridge University Press.
}
\seealso{
The source of the data is
\url{https://feb.kuleuven.be/public/u0043181/modelselection/datasets/egyptlives_data.txt}
}
\keyword{datasets}
