% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get_start}
\alias{get_start}
\title{Returns appropriate starting value}
\usage{
get_start(default, name, ...)
}
\arguments{
\item{default}{Function to calculate default parameter value.}

\item{name}{Name of default starting value.}
}
\value{
Default value if \code{name} is not present
}
\description{
Returns appropriate starting value
}
