% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{inverse_digamma}
\alias{inverse_digamma}
\title{Inverse digamma function}
\usage{
inverse_digamma(y)
}
\arguments{
\item{y}{Values to invert.}
}
\description{
Calculates the inverse digamma function using Newton--Raphson. Works for
y > -500. Uses Newton--Raphson with relative tolerance of eps^0.25.
}
\details{
The number of iterations are few, 1 for most input values, especially those
that are large. The starting value is the lower bound found by Batir (2017).
}
\references{
Batir, N. (2017). Inequalities for the inverses of the polygamma functions. arXiv. http://arxiv.org/abs/1705.06547
}
