% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{MaximumLikelihood}
\alias{MaximumLikelihood}
\alias{dml}
\alias{pml}
\alias{qml}
\alias{rml}
\title{Maximum likelihood estimated distribution}
\usage{
dml(x, obj, log = FALSE)

pml(q = q, obj, lower.tail = TRUE, log.p = FALSE)

qml(p = p, obj, lower.tail = TRUE, log.p = FALSE)

rml(n = n, obj)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{obj}{an \code{univariateML} object.}

\item{log, log.p}{logical; if \code{TRUE}, the probabilities p are gives as
\code{log(p)}.}

\item{lower.tail}{logical; if \code{TRUE} (default), the probabilites are
\eqn{P[X \le x]} otherwise, \eqn{P[X > x]}}

\item{p}{vector of probabilities.}

\item{n}{number of observations. If \code{length(n) > 1}, the length is
taken to be the number required.}
}
\value{
\code{dml} gives the density, \code{pml} gives the distribution
    function, \code{qml} gives the quantile function, and \code{rml}
    generates random deviates.
}
\description{
Density, distribution function, quantile function and random generation for
   a univariate distribution estimated by maximum likelihood.
}
