% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{ml_input_checker}
\alias{ml_input_checker}
\title{Input Checker for ML functions}
\usage{
ml_input_checker(x)
}
\arguments{
\item{x}{input to a `ML***` function.}
}
\value{
`NULL`
}
\description{
Checks that `x` in the ML functions is numeric and has only one dimension.
}
