% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlbeta.R
\name{mlbeta}
\alias{mlbeta}
\title{Estimates the parameter of the Beta distribution using maximum likelihood}
\usage{
mlbeta(x, na.rm = FALSE, start = NULL, type = c("none", "gradient",
  "hessian"))
}
\arguments{
\item{x}{The data from which the estimate is to be computed.}

\item{na.rm}{logical. Should missing values be removed?}

\item{start}{Optional starting parameter values for the minimization.
Passed to the \code{stats::nlm} function.}

\item{type}{Whether a dedicated \code{"gradient"}, \code{"hessian"}, or
\code{"none"} should be passed to \code{stats::nlm}.}
}
\value{
A named numeric vector with maximum likelihood estimates for
 \code{shape1} and \code{shape2}.
}
\description{
Uses \code{stat::nlm} to estimate the parameters of the Beta distribution.
}
\details{
For the density function of the Beta distribution see \link[stats]{dbeta}.

For \code{type}, the option \code{none} is fastest.
}
\references{
Johnson, N. L., Kotz, S. and Balakrishnan, N. (1995) Continuous Univariate Distributions, Volume 2, Chapter 25. Wiley, New York.
}
\seealso{
\link[stats]{dbeta} for the Beta density, \link[stats]{nlm} for the
  optimizer this function uses.
}
