% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlbetapr.R
\name{mlbetapr}
\alias{mlbetapr}
\title{Estimates the parameter of the Beta prime distribution using maximum likelihood}
\usage{
mlbetapr(x, na.rm = TRUE, start = NULL, type = c("none", "gradient",
  "hessian"))
}
\arguments{
\item{x}{The data from which the estimate is to be computed.}

\item{na.rm}{logical. Should missing values be removed?}

\item{start}{Optional starting parameter values for the minimization.
Passed to the \code{stats::nlm} function.}

\item{type}{Whether a dedicated \code{gradient} or \code{hessian} should be
passed to \code{stats::nlm}.}
}
\value{
A named numeric vector with maximum likelihood estimates for
\code{shape1} and \code{shap2}.
}
\description{
Transforms the data and uses \code{stat::nlm} to estimate the parameters
    of the Beta distribution.
}
\details{
For \code{type}, the option \code{none} is fastest.
}
