% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlexp.R
\name{mlexp}
\alias{mlexp}
\title{Estimates the parameter of the exponential distribution using maximum likelihood}
\usage{
mlexp(x, na.rm = FALSE)
}
\arguments{
\item{x}{The data from which the estimate is to be computed.}

\item{na.rm}{logical. Should missing values be removed?}
}
\value{
A named numeric vector with maximum likelihood estimates for
\code{rate}.
}
\description{
The maximum likelihood estimate of \code{rate} is the inverse sample mean.
}
