% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mllogis.R
\name{mllogis}
\alias{mllogis}
\title{Estimates the parameter of the logistic distribution using maximum likelihood}
\usage{
mllogis(x, na.rm = FALSE)
}
\arguments{
\item{x}{The data from which the estimate is to be computed.}

\item{na.rm}{logical. Should missing values be removed?}
}
\value{
A named numeric vector with maximum likelihood estimates for
\code{location} and \code{scale}.
}
\description{
Calculates the estimates using \code{nlm} and an exponential transform of the
    location parameter. If \code{n < 5}, an exact solution is reported. In
    the edge case where no maximum likelihood estimator exists and error is
    thrown.
}
