% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mllogitnorm.R
\name{mllogitnorm}
\alias{mllogitnorm}
\title{Logit-Normal distribution maximum likelihood estimation}
\usage{
mllogitnorm(x, na.rm = FALSE)
}
\arguments{
\item{x}{The data from which the estimate is to be computed.}

\item{na.rm}{logical. Should missing values be removed?}
}
\value{
A named numeric vector with maximum likelihood estimates for
\code{mu} and \code{sigma}.
}
\description{
The maximum likelihood estimate of \code{mu} is the empirical mean of the
    logit transformed data and the maximum likelihood estimate of
    \code{sigma} is the square root of the logit transformed
    biased sample variance.
}
\details{
For the density function of the logit-normal distribution see \link[logitnorm]{dlogitnorm}.
}
\examples{
AIC(mllogitnorm(USArrests$Rape/100))
}
\references{
Atchison, J., & Shen, S. M. (1980). Logistic-normal distributions: Some properties and uses. Biometrika, 67(2), 261-272.
}
\seealso{
link[dlogitnorm]{dlogitnorm}for the normal density.
}
