% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mllomax.R
\name{mllomax}
\alias{mllomax}
\title{Estimates the parameters of the Lomax distribution by maximum likelihood}
\usage{
mllomax(x, na.rm = FALSE, start = 1/stats::median(x),
  rel.tol = .Machine$double.eps^0.25, iterlim = 100)
}
\arguments{
\item{x}{The data from which the estimate is to be computed.}

\item{na.rm}{logical. Should missing values be removed?}

\item{start}{An optional starting value for the \code{lambda} parameter.
Defaults to \code{1/median(x)}.}

\item{rel.tol}{Relative accuracy requested.}

\item{iterlim}{A positive integer specifying the maximum number of
iterations to be performed before the program is terminated.}
}
\value{
A named numeric vector with maximum likelihood estimates for
\code{lambda} and \code{kappa}.
}
\description{
Uses Newton-Raphson to estimate the parameters of the Lomax distribution.
}
