% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlpareto.R
\name{mlpareto}
\alias{mlpareto}
\title{Estimates the parameter of the Pareto distribution using maximum likelihood}
\usage{
mlpareto(x, na.rm = FALSE)
}
\arguments{
\item{x}{The data from which the estimate is to be computed.}

\item{na.rm}{logical. Should missing values be removed?}
}
\value{
A named numeric vector with maximum likelihood estimates for
    \code{a} and \code{b}.
}
\description{
The maximum likelihood estimate of \code{b} is the minimum of \code{x} and the
    maximum likelihood estimate of \code{a} is
    \code{1/(mean(log(x)) - log(b))}.
}
\details{
For the density function of the Pareto distribution see \link[extraDistr]{dpareto}.
}
\examples{
mlpareto(precip)
}
\references{
Johnson, N. L., Kotz, S. and Balakrishnan, N. (1995) Continuous Univariate Distributions, Volume 1, Chapter 20. Wiley, New York.
}
\seealso{
\link[extraDistr]{dpareto} for the Pareto density.
}
