% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlpower.R
\name{mlpower}
\alias{mlpower}
\title{Estimates the parameter of the power distribution using maximum likelihood}
\usage{
mlpower(x, na.rm = FALSE, epsilon = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{The data from which the estimate is to be computed.}

\item{na.rm}{logical. Should missing values be removed?}

\item{epsilon}{Positive number added to \code{max(x)} as an approximation
to the maximum likelihood.}
}
\value{
A named numeric vector with maximum likelihood estimates for
    \code{alpha} and \code{beta}.
}
\description{
Estimates the parameter of the power distribution using maximum likelihood
}
