% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlrayleigh.R
\name{mlrayleigh}
\alias{mlrayleigh}
\title{Estimates the parameter of the Rayleigh distribution using maximum likelihood}
\usage{
mlrayleigh(x, na.rm = FALSE)
}
\arguments{
\item{x}{The data from which the estimate is to be computed.}

\item{na.rm}{logical. Should missing values be removed?}
}
\value{
A named numeric vector with maximum likelihood estimates for
\code{location} and \code{scale}.
}
\description{
Calculates the \code{sigma} parameter as the square root of half the
   empirical second moment.
}
