% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mltnorm.R
\name{mltnorm}
\alias{mltnorm}
\title{Estimates of the underlying mean and standard deviation of a truncated normal}
\usage{
mltnorm(x, a = -Inf, b = Inf, na.rm = FALSE)
}
\arguments{
\item{x}{The data from which the estimate is to be computed.}

\item{a, b}{lower and upper truncation points, \code{a < x <= b}.}

\item{na.rm}{logical. Should missing values be removed?}
}
\value{
A named numeric vector with maximum likelihood estimates for
\code{location} and \code{scale}.
}
\description{
Calculates the maximum likelihood estimates of \code{mean} and \code{sd} of
   a truncated normal with known lower truncation \code{a} and upper t
   truncation \code{b}.
}
