% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlwald.R
\name{mlwald}
\alias{mlwald}
\title{Estimates the parameter of the Wald (inverse Gaussian) distribution using maximum likelihood}
\usage{
mlwald(x, na.rm = FALSE)
}
\arguments{
\item{x}{The data from which the estimate is to be computed.}

\item{na.rm}{logical. Should missing values be removed?}
}
\value{
A named numeric vector with maximum likelihood estimates for
\code{mu} and \code{lambda}.
}
\description{
The maximum likelihood estimate of \code{mu} is the empirical mean and the
    maximum likelihood estimate of \code{lambda} is the difference between
    the mean of reciprocals and the reciprocal of the mean.
}
