% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{newton_raphson_1d}
\alias{newton_raphson_1d}
\title{One-dimensional Newton--Raphson}
\usage{
newton_raphson_1d(f_over_df, param0, ...)
}
\arguments{
\item{f_over_df}{The term f(x_0) / f'(x_0) in Newton--Raphson.}

\item{param0}{Starting parameter, typically from method of moments estimates.}

\item{...}{Optional \code{iterlim} and \code{reltol} parameters.}
}
\value{
Newton--Raphson estimate.
}
\description{
One-dimensional Newton--Raphson
}
\keyword{internal}
