% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{plot_wrangler}
\alias{plot_wrangler}
\title{Wrangles arguments for use in the plot, lines and points functions.}
\usage{
plot_wrangler(x, range, points = FALSE, kind, ...)
}
\arguments{
\item{x}{The input data.}

\item{range}{Range of the data.}

\item{points}{Boolean; should points be plotted by default?}
}
\description{
Wrangles arguments for use in the plot, lines and points functions.
}
\keyword{internal}
