% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{ppqq_wrangler}
\alias{ppqq_wrangler}
\title{Wrangles arguments for use in ppml and qqml functions.}
\usage{
ppqq_wrangler(y, obj, datax, pp, ...)
}
\arguments{
\item{y}{The input data.}

\item{obj}{Function or continuous \code{"univariateML"} object.}

\item{datax}{logical; if true, plots the data on the x axis.}

\item{...}{Arguments passed to \code{plot} or \code{points} down the line.}
}
\description{
Wrangles arguments for use in ppml and qqml functions.
}
\keyword{internal}
