% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{simulate_default}
\alias{simulate_default}
\title{Simulate \code{n} observations from a \verb{ml***} string using default parameters.}
\usage{
simulate_default(dens, n)
}
\arguments{
\item{dens}{\verb{ml***} string.}

\item{n}{Number of samples to take.}
}
\description{
Simulate \code{n} observations from a \verb{ml***} string using default parameters.
}
\keyword{internal}
