% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{to_univariateML}
\alias{to_univariateML}
\title{Data and function to 'univariateML'}
\usage{
to_univariateML(y, obj)
}
\arguments{
\item{y}{Data to transform.}

\item{obj}{Function to apply.}
}
\description{
Data and function to 'univariateML'
}
\keyword{internal}
