% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariateML.R
\docType{package}
\name{univariateML-package}
\alias{univariateML}
\alias{univariateML-package}
\title{univariateML}
\description{
An \code{R}-package for fast, easy, and reliable maximum likelihood estimation for a
selection of parameteric univariate densities.
}
\details{
Data analysis often requieres the estimation of univariate densities. For
almost all univariate densities maximum likelihood estimation is sensible and
gives consistent and efficient estimators. The purpose of this package is to
do maximum for a large number of densities

\itemize{
  \item It supports maximum likelihood estimation of a large selection of densities.
  \item It supports plently of generics such as \code{plot} and \code{AIC} which aids in data analysis.
}

Read the vignettes to learn more about univariateML:
\code{browseVignettes(package = "univariateML")}
}
