% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{univariateML_construct}
\alias{univariateML_construct}
\title{Construct \code{univariateML} object.}
\usage{
univariateML_construct(estimates, name, params)
}
\arguments{
\item{estimates}{The estimated parameters}

\item{name}{Name of the \verb{ml***} function.}

\item{params}{List of \code{loglik}, \code{call}, and \code{n}.}
}
\value{
Object of class \code{univariateML}
}
\description{
Construct \code{univariateML} object.
}
