% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlselect.R
\docType{data}
\name{univariateML_families}
\alias{univariateML_families}
\title{Implemented distribution families}
\format{
An object of class \code{character} of length 24.
}
\usage{
univariateML_families
}
\description{
Implemented distribution families
}
\examples{
print(univariateML_families)
}
\keyword{datasets}
