% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{data}
\name{univariateML_metadata}
\alias{univariateML_metadata}
\title{Metadata for \code{univariateML} models.}
\format{
An object of class \code{list} of length 42.
}
\usage{
univariateML_metadata
}
\description{
Metadata for \code{univariateML} models.
}
\keyword{datasets}
