% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_select.R
\docType{data}
\name{univariateML_models}
\alias{univariateML_models}
\title{Implemented models}
\format{
An object of class \code{character} of length 29.
}
\usage{
univariateML_models
}
\description{
Implemented models
}
\examples{
print(univariateML_models)
}
\keyword{datasets}
