% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{univariateML_to_string}
\alias{univariateML_to_string}
\title{Transform a univariateML object to a string specifying quantile, CDF,
density or random variate generation.}
\usage{
univariateML_to_string(obj, type = c("d", "p", "q", "r", "ml"))
}
\arguments{
\item{obj}{A univariateML object.}

\item{type}{A type.}
}
\value{
A string
}
\description{
Transform a univariateML object to a string specifying quantile, CDF,
density or random variate generation.
}
\keyword{internal}
