% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlbetapr.R
\name{mlbetapr}
\alias{mlbetapr}
\title{Estimates the parameter of the Beta prime distribution using maximum likelihood}
\usage{
mlbetapr(x, na.rm = FALSE, start = NULL, type = c("none", "gradient",
  "hessian"))
}
\arguments{
\item{x}{The data from which the estimate is to be computed.}

\item{na.rm}{logical. Should missing values be removed?}

\item{start}{Optional starting parameter values for the minimization.
Passed to the \code{stats::nlm} function.}

\item{type}{Whether a dedicated \code{gradient} or \code{hessian} should be
passed to \code{stats::nlm}.}
}
\value{
A named numeric vector with maximum likelihood estimates for
\code{shape1} and \code{shap2}.
}
\description{
This function does not estimate the scale parameter for the \code{BetaPrime}
    distribution. Transforms the data and uses \code{stat::nlm} to estimate
    the parameters of the Beta distribution.
}
\details{
For the density function of the Beta prime distribution see \link[extraDistr]{dbetapr}.

For \code{type}, the option \code{none} is fastest.
}
\examples{
AIC(mlbetapr(USArrests$Rape))
}
\references{
Johnson, N. L., Kotz, S. and Balakrishnan, N. (1995) Continuous Univariate Distributions, Volume 2, Chapter 25. Wiley, New York.
}
\seealso{
\link[extraDistr]{dbetapr} for the Beta prime density, \link[stats]{nlm} for the
  optimizer this function uses, \link{mlbeta} for the Beta distribution maximum
  likelihood estimator.
}
