% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlcauchy.R
\name{mlcauchy}
\alias{mlcauchy}
\title{Estimates the parameter of the Cauchy distribution using maximum likelihood}
\usage{
mlcauchy(x, na.rm = FALSE)
}
\arguments{
\item{x}{The data from which the estimate is to be computed.}

\item{na.rm}{logical. Should missing values be removed?}
}
\value{
A named numeric vector with maximum likelihood estimates for
\code{location} and \code{scale}.
}
\description{
Calculates the estimates using \code{nlm} and an exponential transform of the
    location parameter. If \code{n < 5}, an exact solution is reported. In
    the edge case where no maximum likelihood estimator exists and error is
    thrown.
}
\details{
For the density function of the Beta distribution see \link[extraDistr]{dcauchy}.
}
\references{
#' @references Johnson, N. L., Kotz, S. and Balakrishnan, N. (1995) Continuous Univariate Distributions, Volume 1, Chapter 16. Wiley, New York.
}
\seealso{
\link[extraDistr]{dcauchy} for the Cauchy density, \link[stats]{nlm} for the
  optimizer this function uses.
}
