% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlgumbel.R, R/mllgamma.R, R/mlllogis.R,
%   R/mllnorm.R, R/mllogis.R, R/mllogitnorm.R, R/mllomax.R, R/mlnorm.R,
%   R/mlpareto.R, R/mlpois.R, R/mlpower.R, R/mlrayleigh.R, R/mlsged.R,
%   R/mlsnorm.R, R/mlsstd.R, R/mlstd.R, R/mlunif.R, R/mlweibull.R
\name{mlgumbel}
\alias{mlgumbel}
\title{Gumbel distribution maximum likelihood estimation}
\usage{
mlgumbel(x, na.rm = FALSE, ...)

mlgumbel(x, na.rm = FALSE, ...)

mlgumbel(x, na.rm = FALSE, ...)

mlgumbel(x, na.rm = FALSE, ...)

mlgumbel(x, na.rm = FALSE, ...)

mlgumbel(x, na.rm = FALSE, ...)

mlgumbel(x, na.rm = FALSE, ...)

mlgumbel(x, na.rm = FALSE, ...)

mlgumbel(x, na.rm = FALSE, ...)

mlgumbel(x, na.rm = FALSE, ...)

mlgumbel(x, na.rm = FALSE, ...)

mlgumbel(x, na.rm = FALSE, ...)

mlgumbel(x, na.rm = FALSE, ...)

mlgumbel(x, na.rm = FALSE, ...)

mlgumbel(x, na.rm = FALSE, ...)

mlgumbel(x, na.rm = FALSE, ...)

mlgumbel(x, na.rm = FALSE, ...)

mlgumbel(x, na.rm = FALSE, ...)
}
\arguments{
\item{x}{a (non-empty) numeric vector of data values.}

\item{na.rm}{logical. Should missing values be removed?}

\item{...}{\code{shape0} is an optional starting value for the \code{shape} parameter.
\code{rel.tol} is the relative accuracy requested, defaults
to \code{.Machine$double.eps^0.25}. \code{iterlim} is a positive integer
specifying the maximum number of iterations to be performed before the
program is terminated (defaults to \code{100}).}
}
\value{
\code{mlgumbel} returns an object of \link[base:class]{class} \code{univariateML}.
This is a named numeric vector with maximum likelihood estimates for \code{mu}
and \code{s} and the following attributes:
\item{\code{model}}{The name of the model.}
\item{\code{density}}{The density associated with the estimates.}
\item{\code{logLik}}{The loglikelihood at the maximum.}
\item{\code{support}}{The support of the density.}
\item{\code{n}}{The number of observations.}
\item{\code{call}}{The call as captured my \code{match.call}}
\code{shape} and \code{sigma}.

\code{mllgamma} returns an object of \link[base:class]{class} \code{univariateML}.
This is a named numeric vector with maximum likelihood estimates for
\code{shapelog} and \code{ratelog} and the following attributes:
\item{\code{model}}{The name of the model.}
\item{\code{density}}{The density associated with the estimates.}
\item{\code{logLik}}{The loglikelihood at the maximum.}
\item{\code{support}}{The support of the density.}
\item{\code{n}}{The number of observations.}
\item{\code{call}}{The call as captured my \code{match.call}}

\code{mlllogis} returns an object of \link[base:class]{class} \code{univariateML}.
This is a named numeric vector with maximum likelihood estimates for
\code{shape} and \code{rate} and the following attributes:
\item{\code{model}}{The name of the model.}
\item{\code{density}}{The density associated with the estimates.}
\item{\code{logLik}}{The loglikelihood at the maximum.}
\item{\code{support}}{The support of the density.}
\item{\code{n}}{The number of observations.}
\item{\code{call}}{The call as captured my \code{match.call}}

\code{mllnorm} returns an object of \link[base:class]{class} \code{univariateML}.
This is a named numeric vector with maximum likelihood estimates for
\code{meanlog} and \code{sdlog} and the following attributes:
\item{\code{model}}{The name of the model.}
\item{\code{density}}{The density associated with the estimates.}
\item{\code{logLik}}{The loglikelihood at the maximum.}
\item{\code{support}}{The support of the density.}
\item{\code{n}}{The number of observations.}
\item{\code{call}}{The call as captured my \code{match.call}}

\code{mllogis} returns an object of \link[base:class]{class} \code{univariateML}.
This is a named numeric vector with maximum likelihood estimates for
\code{location} and \code{scale} and the following attributes:
\item{\code{model}}{The name of the model.}
\item{\code{density}}{The density associated with the estimates.}
\item{\code{logLik}}{The loglikelihood at the maximum.}
\item{\code{support}}{The support of the density.}
\item{\code{n}}{The number of observations.}
\item{\code{call}}{The call as captured my \code{match.call}}

\code{mllogitnorm} returns an object of \link[base:class]{class}
\code{univariateML}. This is a named numeric vector with maximum likelihood
estimates for \code{mu} and \code{sigma} and the following attributes:
\item{\code{model}}{The name of the model.}
\item{\code{density}}{The density associated with the estimates.}
\item{\code{logLik}}{The loglikelihood at the maximum.}
\item{\code{support}}{The support of the density.}
\item{\code{n}}{The number of observations.}
\item{\code{call}}{The call as captured my \code{match.call}}

\code{mllomax} returns an object of \link[base:class]{class} \code{univariateML}.
This is a named numeric vector with maximum likelihood estimates for
\code{lambda} and \code{kappa} and the following attributes:
\item{\code{model}}{The name of the model.}
\item{\code{density}}{The density associated with the estimates.}
\item{\code{logLik}}{The loglikelihood at the maximum.}
\item{\code{support}}{The support of the density.}
\item{\code{n}}{The number of observations.}
\item{\code{call}}{The call as captured my \code{match.call}}

\code{mlnorm} returns an object of \link[base:class]{class} \code{univariateML}.
This is a named numeric vector with maximum likelihood estimates for
\code{mean} and \code{sd} and the following attributes:
\item{\code{model}}{The name of the model.}
\item{\code{density}}{The density associated with the estimates.}
\item{\code{logLik}}{The loglikelihood at the maximum.}
\item{\code{support}}{The support of the density.}
\item{\code{n}}{The number of observations.}
\item{\code{call}}{The call as captured my \code{match.call}}

\code{mlpareto} returns an object of \link[base:class]{class} \code{univariateML}.
This is a named numeric vector with maximum likelihood estimates for
\code{a} and \code{b} and the following attributes:
\item{\code{model}}{The name of the model.}
\item{\code{density}}{The density associated with the estimates.}
\item{\code{logLik}}{The loglikelihood at the maximum.}
\item{\code{support}}{The support of the density.}
\item{\code{n}}{The number of observations.}
\item{\code{call}}{The call as captured my \code{match.call}}

\code{mlpois} returns an object of \link[base:class]{class} \code{univariateML}.
This is a named numeric vector with maximum likelihood estimates for
\code{mean} and \code{sd} and the following attributes:
\item{\code{model}}{The name of the model.}
\item{\code{density}}{The density associated with the estimates.}
\item{\code{logLik}}{The loglikelihood at the maximum.}
\item{\code{support}}{The support of the density.}
\item{\code{n}}{The number of observations.}
\item{\code{call}}{The call as captured my \code{match.call}}

\code{mlpower} returns an object of \link[base:class]{class} \code{univariateML}.
This is a named numeric vector with maximum likelihood estimates for
\code{alpha} and \code{beta} and the following attributes:
\item{\code{model}}{The name of the model.}
\item{\code{density}}{The density associated with the estimates.}
\item{\code{logLik}}{The loglikelihood at the maximum.}
\item{\code{support}}{The support of the density.}
\item{\code{n}}{The number of observations.}
\item{\code{call}}{The call as captured my \code{match.call}}

\code{mlrayleigh} returns an object of \link[base:class]{class}
\code{univariateML}. This is a named numeric vector with maximum likelihood
estimates for \code{sigma} and the following attributes:
\item{\code{model}}{The name of the model.}
\item{\code{density}}{The density associated with the estimates.}
\item{\code{logLik}}{The loglikelihood at the maximum.}
\item{\code{support}}{The support of the density.}
\item{\code{n}}{The number of observations.}
\item{\code{call}}{The call as captured my \code{match.call}}

\code{mlsged} returns an object of \link[base:class]{class} \code{univariateML}.
This is a named numeric vector with maximum likelihood estimates for the
parameters \code{mean}, \code{sd}, \code{nu}, \code{xi}, and the following attributes:
\item{\code{model}}{The name of the model.}
\item{\code{density}}{The density associated with the estimates.}
\item{\code{logLik}}{The loglikelihood at the maximum.}
\item{\code{support}}{The support of the density.}
\item{\code{n}}{The number of observations.}
\item{\code{call}}{The call as captured my \code{match.call}}

\code{mlsnorm} returns an object of \link[base:class]{class} \code{univariateML}.
This is a named numeric vector with maximum likelihood estimates for
the parameters \code{mean}, \code{sd}, \code{xi} and the following attributes:
\item{\code{model}}{The name of the model.}
\item{\code{density}}{The density associated with the estimates.}
\item{\code{logLik}}{The loglikelihood at the maximum.}
\item{\code{support}}{The support of the density.}
\item{\code{n}}{The number of observations.}
\item{\code{call}}{The call as captured my \code{match.call}}

\code{mlsstd} returns an object of \link[base:class]{class} \code{univariateML}.
This is a named numeric vector with maximum likelihood estimates for
the parameters \code{mean}, \code{sd}, \code{nu}, \code{xi} and the following attributes:
\item{\code{model}}{The name of the model.}
\item{\code{density}}{The density associated with the estimates.}
\item{\code{logLik}}{The loglikelihood at the maximum.}
\item{\code{support}}{The support of the density.}
\item{\code{n}}{The number of observations.}
\item{\code{call}}{The call as captured my \code{match.call}}

\code{mlstd} returns an object of \link[base:class]{class} \code{univariateML}.
This is a named numeric vector with maximum likelihood estimates for
the parameters \code{mean}, \code{sd}, \code{nu} and the following attributes:
\item{\code{model}}{The name of the model.}
\item{\code{density}}{The density associated with the estimates.}
\item{\code{logLik}}{The loglikelihood at the maximum.}
\item{\code{support}}{The support of the density.}
\item{\code{n}}{The number of observations.}
\item{\code{call}}{The call as captured my \code{match.call}}

\code{mlunif} returns an object of \link[base:class]{class} \code{univariateML}.
This is a named numeric vector with maximum likelihood estimates for \code{min}
and \code{max} and the following attributes:
\item{\code{model}}{The name of the model.}
\item{\code{density}}{The density associated with the estimates.}
\item{\code{logLik}}{The loglikelihood at the maximum.}
\item{\code{support}}{The support of the density.}
\item{\code{n}}{The number of observations.}
\item{\code{call}}{The call as captured my \code{match.call}}

\code{mlweibull} returns an object of \link[base:class]{class} \code{univariateML}.
This is a named numeric vector with maximum likelihood estimates for
\code{shape} and \code{scale} and the following attributes:
\item{\code{model}}{The name of the model.}
\item{\code{density}}{The density associated with the estimates.}
\item{\code{logLik}}{The loglikelihood at the maximum.}
\item{\code{support}}{The support of the density.}
\item{\code{n}}{The number of observations.}
\item{\code{call}}{The call as captured my \code{match.call}}
}
\description{
Uses Newton-Raphson to estimate the parameters of the Gumbel distribution.

The maximum likelihood estimate of \code{shapelog} and \code{ratelog} are calculated
by calling \code{\link[=mlgamma]{mlgamma()}} on the transformed data.

The maximum likelihood estimate of \code{shape} and \code{rate} are calculated
by transforming the data back to the logistic model and applying
\code{\link{mllogis}}.

The maximum likelihood estimate of \code{meanlog} is the empirical mean of the
log-transformed data and the maximum likelihood estimate of \code{sdlog}
is the square root of the biased sample variance based on the
log-transformed data.

Calculates the estimates using \code{nlm} with an exponential transform of the
location parameter.

The maximum likelihood estimate of \code{mu} is the empirical mean of the
logit transformed data and the maximum likelihood estimate of
\code{sigma} is the square root of the logit transformed
biased sample variance.

Uses Newton-Raphson to estimate the parameters of the Lomax distribution.

The maximum likelihood estimate of \code{mean} is the empirical mean and the
maximum likelihood estimate of \code{sd} is the square root of the
biased sample variance.

The maximum likelihood estimate of \code{b} is the minimum of \code{x} and the
maximum likelihood estimate of \code{a} is
\code{1/(mean(log(x)) - log(b))}.

The maximum likelihood estimate of \code{lambda} is the empirical mean.

The maximum likelihood estimate of \code{alpha} is the maximum of \code{x} +
\code{epsilon} (see the details) and the maximum likelihood estimate of
\code{beta} is \code{1/(log(alpha)-mean(log(x)))}.

Calculates the \code{sigma} parameter as the square root of half the
empirical second moment.

Joint maximum likelihood estimation as implemented by \link[fGarch:dist-sgedFit]{fGarch::sgedFit}.

Joint maximum likelihood estimation as implemented by \link[fGarch:dist-snormFit]{fGarch::snormFit}.

Joint maximum likelihood estimation as implemented by \link[fGarch:dist-sstdFit]{fGarch::sstdFit}.

Joint maximum likelihood estimation as implemented by \link[fGarch:dist-stdFit]{fGarch::stdFit}.

The estimates are \code{min(x)} and \code{max(x)}.

Uses Newton-Raphson to estimate the parameters of the Weibull distribution.
}
\details{
For the density function of the Gumbel distribution see
\link[extraDistr:Gumbel]{Gumbel}.

For the density function of the log normal distribution see
\link[actuar:Loggamma]{Loggamma}.

For the density function of the log-logistic distribution see
\link[actuar:Loglogistic]{Loglogistic}

For the density function of the log normal distribution see
\link[stats:Lognormal]{Lognormal}.

For the density function of the logistic distribution see
\link[stats:Logistic]{Logistic}.

For the density function of the logit-normal distribution see
\link[logitnorm:dlogitnorm]{dlogitnorm}.

For the density function of the Lomax distribution see
\link[extraDistr:Lomax]{Lomax}. The maximum likelihood estimate will frequently
fail to exist. This is due to the parameterization of the function which
does not take into account that the density converges to an exponential
along certain values of the parameters, see
\code{vignette("Distribution Details", package = "univariateML")}.

For the density function of the normal distribution see
\link[stats:Normal]{Normal}.

For the density function of the Pareto distribution see
\link[extraDistr:Pareto]{Pareto}.

For the density function of the Poisson distribution see
\link[stats:Poisson]{Poisson}.

For the density function of the power distribution see
\link[extraDistr:PowerDist]{PowerDist}. The maximum likelihood estimator of
\code{alpha} does not exist, strictly
speaking. This is because \code{x} is supported \code{c(0, alpha)} with
an open endpoint on alpha in the \code{extraDistr} implementation of
\code{dpower}. If the endpoint was closed, \code{max(x)} would have been
the maximum likelihood estimator. To overcome this problem, we add
a possibly user specified \code{epsilon} to \code{max(x)}.

For the density function of the Rayleigh distribution see
\link[extraDistr:Rayleigh]{Rayleigh}.

For the density function of the Student t-distribution see
\link[fGarch:dist-sged]{sged}.

For the density function of the Student t distribution see
\link[fGarch:dist-snorm]{dsnorm}.

For the density function of the skew Student t-distribution see
\link[fGarch:dist-sstd]{sstd}.

For the density function of the Student t-distribution see
\link[fGarch:dist-std]{std}.

For the density function of the logistic distribution see
\link[stats:Uniform]{Uniform}.

For the density function of the Weibull distribution see
\link[stats:Weibull]{Weibull}.
}
\examples{
mlgumbel(precip)
mllgamma(precip)
mllnorm(precip)
mllnorm(precip)
mllogis(precip)
AIC(mllogitnorm(USArrests$Rape / 100))
set.seed(3)
mllomax(extraDistr::rlomax(100, 2, 4))
mlnorm(precip)
mlpareto(precip)
mlpois(precip)
mlpower(precip)
mlrayleigh(precip)
mlsged(precip)
mlsnorm(precip)
mlsstd(precip)
mlstd(precip)
mlunif(precip)
BIC(mlweibull(precip))
}
\references{
Johnson, N. L., Kotz, S. and Balakrishnan, N. (1995) Continuous Univariate Distributions, Volume 2, Chapter 22. Wiley, New York.

Hogg, R. V. and Klugman, S. A. (1984), Loss Distributions, Wiley.

Dutang, C., Goulet, V., & Pigeon, M. (2008). actuar: An R package for
actuarial science. Journal of Statistical Software, 25(7), 1-37.

Kleiber, C. and Kotz, S. (2003), Statistical Size Distributions
in Economics and Actuarial Sciences, Wiley.

Klugman, S. A., Panjer, H. H. and Willmot, G. E. (2012), Loss Models,
From Data to Decisions, Fourth Edition, Wiley.

Dutang, C., Goulet, V., & Pigeon, M. (2008). actuar: An R package for
actuarial science. Journal of Statistical Software, 25(7), 1-37.

Johnson, N. L., Kotz, S. and Balakrishnan, N. (1995)
Continuous Univariate Distributions, Volume 1, Chapter 14. Wiley, New York.

Johnson, N. L., Kotz, S. and Balakrishnan, N. (1995)
Continuous Univariate Distributions, Volume 2, Chapter 23. Wiley, New York.

Atchison, J., & Shen, S. M. (1980). Logistic-normal
distributions: Some properties and uses. Biometrika, 67(2), 261-272.

Kleiber, Christian; Kotz, Samuel (2003), Statistical Size
Distributions in Economics and Actuarial Sciences, Wiley Series in
Probability and Statistics, 470, John Wiley & Sons, p. 60

Johnson, N. L., Kotz, S. and Balakrishnan, N. (1995)
Continuous Univariate Distributions, Volume 1, Chapter 13. Wiley, New York.

Johnson, N. L., Kotz, S. and Balakrishnan, N. (1995) Continuous
Univariate Distributions, Volume 1, Chapter 20. Wiley, New York.

Arslan, G. "A new characterization of the power distribution."
Journal of Computational and Applied Mathematics 260 (2014): 99-102.

Johnson, N. L., Kotz, S. and Balakrishnan, N. (1995) Continuous
Univariate Distributions, Volume 1, Chapter 18. Wiley, New York.

Nelson D.B. (1991); Conditional Heteroscedasticity in Asset
Returns: A New Approach, Econometrica, 59, 347<U+2013>370.

Fernandez C., Steel M.F.J. (2000); On Bayesian Modelling of Fat Tails and
Skewness, Preprint.

Fernandez C., Steel M.F.J. (2000); On Bayesian Modelling of Fat
Tails and Skewness, Preprint.

Fernandez C., Steel M.F.J. (2000); On Bayesian Modelling of Fat
Tails and Skewness, Preprint.

Johnson, N. L., Kotz, S. and Balakrishnan, N. (1995)
Continuous Univariate Distributions, Volume 1, Chapter 13. Wiley, New York.

Johnson, N. L., Kotz, S. and Balakrishnan, N. (1995)
Continuous Univariate Distributions, Volume 2, Chapter 26. Wiley, New York.

Johnson, N. L., Kotz, S. and Balakrishnan, N. (1995) Continuous
Univariate Distributions, Volume 1, Chapter 21. Wiley, New York.
}
\seealso{
\link[extraDistr:Gumbel]{Gumbel} for the Gumbel density.

\link[actuar:Loggamma]{Loggamma} for the log normal density.

\link[actuar:Loglogistic]{Loglogistic} for the log-logistic density.

\link[stats:Lognormal]{Lognormal} for the log normal density.

\link[stats:Logistic]{Logistic} for the Logistic density,
\link[stats:nlm]{nlm} for the optimizer this function uses.

\link[stats:Normal]{Normal} for the normal density.

\link[extraDistr:Lomax]{Lomax} for the Lomax density.

\link[stats:Normal]{Normal} for the normal density.

\link[extraDistr:Pareto]{Pareto} for the Pareto density.

\link[stats:Poisson]{Poisson} for the Poisson density.

\link[extraDistr:PowerDist]{PowerDist} for the power density. \link{Pareto}
for the closely related Pareto distribution.

\link[extraDistr:Rayleigh]{Rayleigh} for the Rayleigh density.

\link[fGarch:dist-sged]{sged} for the Student t-density.

\link[fGarch:dist-snorm]{dsnorm} for the Student-t density.

\link[fGarch:dist-sstd]{sstd} for the Skew Student t-density.

\link[fGarch:dist-std]{std} for the Student-t density.

\link[stats:Uniform]{Uniform} for the uniform density.

\link[stats:Weibull]{Weibull} for the Weibull density.
}
