% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlgumbel.R
\name{mlgumbel}
\alias{mlgumbel}
\title{Estimates the parameter of a Gumbel distribution by maximum likelihood}
\usage{
mlgumbel(x, na.rm = FALSE, sigma0 = 1,
  rel.tol = .Machine$double.eps^0.25, iterlim = 100)
}
\arguments{
\item{x}{The data from which the estimate is to be computed.}

\item{na.rm}{logical. Should missing values be removed?}

\item{sigma0}{An optional starting value for the \code{sigma} parameter.}

\item{rel.tol}{Relative accuracy requested.}

\item{iterlim}{A positive integer specifying the maximum number of
iterations to be performed before the program is terminated.}
}
\value{
A named numeric vector with maximum likelihood estimates for
\code{shape} and \code{sigma}.
}
\description{
Uses Newton-Raphson to estimate the parameters of the Gumbel distribution.
}
\details{
For the density function of the Gumbel distribution see \link[extraDistr]{dgumbel}.
}
\examples{
mlgumbel(precip)
}
\references{
Johnson, N. L., Kotz, S. and Balakrishnan, N. (1995) Continuous Univariate Distributions, Volume 2, Chapter 22. Wiley, New York.
}
\seealso{
\link[extraDistr]{dgumbel} for the Gumbel density.
}
