% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlinvweibull.R
\name{mlinvweibull}
\alias{mlinvweibull}
\title{Inverse Weibull distribution maximum likelihood estimation}
\usage{
mlinvweibull(x, na.rm = FALSE)
}
\arguments{
\item{x}{The data from which the estimate is to be computed.}

\item{na.rm}{logical. Should missing values be removed?}
}
\value{
A named numeric vector with maximum likelihood estimates for
\code{shape} and \code{rate}.
}
\description{
The maximum likelihood estimate of \code{shape} and \code{rate} are calculated
  by calling \code{\link{mlweibull}} on the transformed data.
}
\details{
For the density function of the log normal distribution see \link[actuar]{InverseWeibull}.
}
\examples{
mlinvweibull(precip)
}
\references{
Kleiber, C. and Kotz, S. (2003), Statistical Size Distributions in Economics and Actuarial Sciences, Wiley.

Klugman, S. A., Panjer, H. H. and Willmot, G. E. (2012), Loss Models, From Data to Decisions, Fourth Edition, Wiley.

Dutang, C., Goulet, V., & Pigeon, M. (2008). actuar: An R package for actuarial science. Journal of Statistical Software, 25(7), 1-37.
}
\seealso{
\link[actuar]{InverseWeibull} for the Inverse Weibull density.
}
