% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mllaplace.R
\name{mllaplace}
\alias{mllaplace}
\title{Estimates the parameter of the Laplace distribution using maximum likelihood}
\usage{
mllaplace(x, na.rm = FALSE)
}
\arguments{
\item{x}{The data from which the estimate is to be computed.}

\item{na.rm}{logical. Should missing values be removed?}
}
\value{
A named numeric vector with maximum likelihood estimates for
\code{mu} and \code{sigma}.
}
\description{
The maximum likelihood estimate of \code{mu} is the sample median while the
   maximum likelihood estimate of \code{sigma} is mean absolute deviation
   from the median.
}
\details{
For the density function of the Laplace distribution see \link[extraDistr]{dlaplace}.
}
\examples{
mllaplace(precip)
}
\references{
Johnson, N. L., Kotz, S. and Balakrishnan, N. (1995) Continuous Univariate Distributions, Volume 2, Chapter 24. Wiley, New York.
}
\seealso{
\link[extraDistr]{dlaplace} for the Laplace density.
}
