% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlrayleigh.R
\name{mlrayleigh}
\alias{mlrayleigh}
\title{Estimates the parameter of the Rayleigh distribution using maximum likelihood}
\usage{
mlrayleigh(x, na.rm = FALSE)
}
\arguments{
\item{x}{The data from which the estimate is to be computed.}

\item{na.rm}{logical. Should missing values be removed?}
}
\value{
A named numeric vector with the maximum likelihood estimate for
   \code{sigma}.
}
\description{
Calculates the \code{sigma} parameter as the square root of half the
   empirical second moment.
}
\details{
For the density function of the Rayleigh distribution see \link[extraDistr]{drayleigh}.
}
\examples{
mlrayleigh(precip)
}
\references{
Johnson, N. L., Kotz, S. and Balakrishnan, N. (1995) Continuous Univariate Distributions, Volume 1, Chapter 18. Wiley, New York.
}
\seealso{
\link[extraDistr]{drayleigh} for the Rayleigh density.
}
