% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlunif.R
\name{mlunif}
\alias{mlunif}
\title{Uniform distribution maximum likelihood estimation}
\usage{
mlunif(x, na.rm = FALSE)
}
\arguments{
\item{x}{The data from which the estimate is to be computed.}

\item{na.rm}{logical. Should missing values be removed?}
}
\value{
A named numeric vector with maximum likelihood estimates for
\code{min} and \code{max}.
}
\description{
The estimates are \code{min(x)} and \code{max(x)}.
}
\details{
For the density function of the logistic distribution see \link[stats]{Uniform}.
}
\examples{
mlunif(precip)
}
\references{
Johnson, N. L., Kotz, S. and Balakrishnan, N. (1995) Continuous Univariate Distributions, Volume 2, Chapter 26. Wiley, New York.
}
\seealso{
\link[stats]{Uniform} for the uniform density.
}
