% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlunif.R
\name{mlunif}
\alias{mlunif}
\title{Estimates the parameters of uniform density using maximum likelihood}
\usage{
mlunif(x, na.rm = TRUE)
}
\arguments{
\item{x}{The data from which the estimate is to be computed.}

\item{na.rm}{logical. Should missing values be removed?}
}
\value{
A named numeric vector with maximum likelihood estimates for
\code{min} and \code{max}.
}
\description{
The estimates are \code{min(x)} and \code{max(x)}.
}
