% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlzipf.R
\name{mlzipf}
\alias{mlzipf}
\title{Zipf distribution maximum likelihood estimation}
\usage{
mlzipf(x, na.rm = FALSE, ...)
}
\arguments{
\item{x}{a (non-empty) numeric vector of data values.}

\item{na.rm}{logical. Should missing values be removed?}

\item{...}{Not currently in use.}
}
\value{
\code{mlzipf} returns an object of \link[base:class]{class} \code{univariateML}.
This is a named numeric vector with maximum likelihood estimates for
\code{mean} and \code{sd} and the following attributes:
\item{\code{model}}{The name of the model.}
\item{\code{density}}{The density associated with the estimates.}
\item{\code{logLik}}{The loglikelihood at the maximum.}
\item{\code{support}}{The support of the density.}
\item{\code{n}}{The number of observations.}
\item{\code{call}}{The call as captured my \code{match.call}}
}
\description{
For the density function of the Zipf distribution see
\link[sads:dzipf]{Zipf}.
}
\examples{
mlzipf(precip)
}
\seealso{
\link[sads:dzipf]{Zipf} for the density.
}
