% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariateML.R
\docType{package}
\name{univariateML-package}
\alias{univariateML}
\alias{univariateML-package}
\title{univariateML}
\description{
An \code{R}-package for fast, easy, and reliable maximum likelihood estimation for a
selection of parameteric univariate densities.
}
\details{
Data analysis often requieres the estimation of univariate densities. Maximum likelihood estimation
is sensible for almost every univariate density used in practice. Moreover,
the maximum likelihood estimator is typically consistent and efficient.

The purpose of this package is to
\itemize{
  \item Spport maximum likelihood estimation of a large selection of densities.
  \item Supports plently of generics such as \code{plot} and \code{AIC} to aid your data analysis.
}

Read the vignettes to learn more about univariateML:
\code{browseVignettes(package = "univariateML")}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/JonasMoss/univariateML}
  \item \url{https://univariateml.netlify.com/}
  \item Report bugs at \url{https://github.com/JonasMoss/univariateML/issues}
}

}
\author{
\strong{Maintainer}: Jonas Moss \email{jonas.gjertsen@gmail.com} (0000-0002-6876-6964)

}
