% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densities.R
\docType{data}
\name{univariateML_models}
\alias{univariateML_models}
\title{Implemented models}
\format{
An object of class \code{character} of length 29.
}
\usage{
univariateML_models
}
\description{
Vector of all supported models in \code{univariateML}.
}
\details{
The currently supported models are \link{mlbeta}, \link{mlbetapr}, \link{mlcauchy}, \link{mlexp}, \link{mlgamma}, \link{mlged}, \link{mlgumbel}, \link{mlinvgamma}, \link{mlinvgauss}, \link{mlinvweibull}, \link{mlkumar}, \link{mllaplace}, \link{mllgamma}, \link{mlllogis}, \link{mllnorm}, \link{mllogis}, \link{mllogitnorm}, \link{mllomax}, \link{mlnaka}, \link{mlnorm}, \link{mlpareto}, \link{mlpower}, \link{mlrayleigh}, \link{mlsged}, \link{mlsnorm}, \link{mlsstd}, \link{mlstd}, \link{mlunif}, \link{mlweibull}
}
\keyword{datasets}
